requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
scikit-learn>=1.3
vaderSentiment>=3.3
requests>=2.28
openpyxl>=3.1
xlrd>=2.0
python-docx>=1.1
streamlit>=1.28
jupyter>=1.0
nbconvert>=7.0
nbclient>=0.7
